+++ applications/vessel-history/src/App.test.tsx
@@ -1,18 +1,14 @@
 import { Provider } from 'react-redux'
-import React from 'react'
-import { render } from '@testing-library/react'
-import useGFWLogin from '@globalfishingwatch/react-hooks/dist/use-login'
+import { render, waitFor } from '@testing-library/react'
+import { useUser } from 'features/user/user.hooks'
 import store from './store'
 import App from './App'
 
-jest.mock('@globalfishingwatch/react-hooks/dist/use-login')
+jest.mock('features/user/user.hooks')
 jest.useFakeTimers()
 
 describe('<App />', () => {
-  const assignMock = jest.fn()
-  const mockGFWLogin: jest.Mock = useGFWLogin as jest.Mock
-  delete window['location']
-  window.location = { assign: assignMock }
+  const mockUser: jest.Mock = useUser as jest.Mock
 
   const gfwLoginDefault = {
     loading: true,
@@ -25,29 +21,20 @@ describe('<App />', () => {
     jest.clearAllMocks()
   })
 
-  it('renders splash screen while loading', () => {
-    mockGFWLogin.mockReturnValue(gfwLoginDefault)
+  it('renders splash screen while loading', async () => {
+    mockUser.mockReturnValue(gfwLoginDefault)
     const component = render(
       <Provider store={store}>
         <App />
       </Provider>
     )
-
-    expect(setTimeout).toHaveBeenLastCalledWith(expect.any(Function), 1000)
+    const splashElement = await waitFor(() => component.getByTestId('splash'))
+    expect(splashElement).toBeInTheDocument()
     expect(component.asFragment()).toMatchSnapshot()
   })
 
-  it('redirects to login screen when not logged', () => {
-    mockGFWLogin.mockReturnValue({ ...gfwLoginDefault, loading: false })
-    render(
-      <Provider store={store}>
-        <App />
-      </Provider>
-    )
-    expect(assignMock).toHaveBeenCalled()
-  })
   it('renders home screen when not loading', () => {
-    mockGFWLogin.mockReturnValue({
+    mockUser.mockReturnValue({
       ...gfwLoginDefault,
       loading: false,
       logged: true,